refactor: split an observer mod from subscribable mod
